improved makefile

+++ Makefile
@@ -1,7 +1,7 @@
 CC ?= cc
 C_FLAGS := -std=c99 $\
 					 -Wall -Wextra -Wpedantic $\
-					 -Og -g -march=native -pipe $\
+					 -O2 -march=native -pipe $\
 					 -Iinclude
 
 DIRECTORIES := build build/tests
@@ -9,10 +9,15 @@ DIRECTORIES := build build/tests
 OBJECT_FILES := build/internal.o build/manipulation.o build/memory.o
 TEST_OBJECT_FILES := build/tests/main.o
 
-all: ${DIRECTORIES} liborchestra.a test
+all: ${DIRECTORIES} liborchestra.a
 
 ${DIRECTORIES}:
-	-mkdir ${DIRECTORIES}
+	$(foreach DIRECTORY,$\
+		${DIRECTORIES},$\
+		$(if $(wildcard ${DIRECTORY}),,$\
+			$(shell mkdir ${DIRECTORY})$\
+		)$\
+	)
 
 build/%.o: src/%.c
 	${CC} -c $< ${C_FLAGS} -o $@
@@ -24,5 +29,14 @@ test: liborchestra.a ${TEST_OBJECT_FILES}
 	${CC} ${TEST_OBJECT_FILES} -L. -lorchestra -o test
 
 clean:
-	-rm -rf ${DIRECTORIES}
-	-rm -f liborchestra.a
+	$(foreach DIRECTORY,$\
+		${DIRECTORIES},$\
+		$(if $(wildcard ${DIRECTORY}),$\
+			$(shell rm -rf ${DIRECTORY})$\
+		)$\
+	)
+ifneq (, $(wildcard liborchestra.a))
+	rm -f liborchestra.a
+endif
+
+.PHONY: all clean
